Handle configurations with unset SSIDs
This was found in the wild where the SSID was populated from the
Application environment. The setting wasn't in the application
environment so `nil` was being passed. This would cause a crash (since
it's wrong) and remove all WiFi configuration. This seems too harsh. The
new behavior is to drop `nil` networks and show a warning.