Add quick_configure helpers
Run `VintageNetWiFi.quick_configure("my_ssid", "password")` and it will
create a WPA PSK WiFi config and apply it to "wlan0". If a password
isn't supplied, it will configure "wlan0" to connect to an open WiFi
access point.

This function is intended to make setting up WiFi really easy for the
majority of users. Anything outside of the norm should call
`VintageNet.configure/3` as usual.