Add Vagrant config for testing WiFi
This config makes it possible to simulate various WiFi configurations so
that we don't need to test on real hardware.

It's in a VM since it uses the `mac80211_hwsim` kernel driver and it's
important that its network interfaces are not touched by anything on the
host system.  This should also be possible using Docker, but I'm not
familiar with how to set that up as nicely as with Vagrant.