Serialize wpa_supplicant requests at the low level
Serializion of messages to the wpa_supplicant was unintentionally done
at a higher level so this doesn't change anything, but its much easier
to see now.

The reason for this change is that I ran an experiment that exercised
multiple outstanding requests without serialization and the dropped
BSSID issue was able to cause responses to be mixed up. I didn't like
this at all since if the upper layers ever were able to schedule
multiple requests simultaneously, the result would almost certainly be
hard to debug.