Fix warnings found by Elixir 1.12
Fixes:

```
warning: module attribute @impl was not set for function ioctl/3 callback (specified in VintageNet.Technology). This either means you forgot to add the "@impl true" annotation before the definition or that you are accidentally overriding this callback
  lib/vintage_net/technology/wifi.ex:35: VintageNet.Technology.WiFi (module)

warning: module attribute @impl was not set for function check_system/1 callback (specified in VintageNet.Technology). This either means you forgot to add the "@impl true" annotation before the definition or that you are accidentally overriding this callback
  lib/vintage_net/technology/wifi.ex:37: VintageNet.Technology.WiFi (module)
```