Parse all wpa_supplicant flags
This adds a new flag parser based on looking at the wpa_supplicant code
base for all of the options. The old way of parsing flags still returns
the same atoms so that this shouldn't break any code. However, the
proto-key_mgmt-cipher flags are now broken out to avoid having to add
every variant that appears in the wild.