Raise command timeout to wpa_supplicant
This raises the timeout from 1 second to 4 seconds. The GenServer
timeout is still 5 seconds, so there's still some headroom before timing
out with an exception.

The reason for this change is 1 second is too short for the initial
connection to the `wpa_supplicant` on a GRiSP 2. While the short timeout
still works, it causes the `wpa_supplicant` to be restarted when it
would have been faster to just wait for the slightly late response.