Filter out mesh peers when updating all access points
When VintagNetWiFi scan would request to get all access points and the
callback function to update all access points would run, the mesh peers
would not be filtered out.

I updated the generic scan test to include a mesh peer and ensured the
update all access point callback function was ran. This allows the code
to be exercised.