Make generic configuration backwards compatible
This makes it so that if you use the generic WiFi configuration and
revert firmware to a previous version of VintageNetWiFi, you'll get a
WPA2 PSK configuration.

This also sets up a future update when v0.12.0 is an old memory to
deprecate `:allowed_key_mgmt` and just use `:key_mgmt`. This version
is accommodating to configuration inputs.