Make management frame protection optional in generic config
The generic configuration was requiring frame protection (ieee80211w=2)
since WPA3 requires it. A lot of WPA2 networks support it so this worked
for those as well. The Samsung S23 does not support it in hotspot mode,
though, iee80211w has to be set to optional (1) to support it.

This has also been tested with various WPA2 WiFi networks that supported
frame protection and a WPA3 SAE network.